requirements.txt
oracledb
psycopg2-binary
tenacity

